+++ src/z80asm/z80asm2/location.h
@@ -12,15 +12,15 @@
 
 class Location {
 public:
-    Location(const std::string& filename = "", int line_num = 0);
+    Location(const std::string& filename = "", size_t line_num = 0);
     void clear();
     bool empty() const;
 
     const std::string& filename() const;
-    int line_num() const;
+    size_t line_num() const;
 
     void set_filename(const std::string& filename);
-    void set_line_num(int line_num);
+    void set_line_num(size_t line_num);
     void inc_line_num();
 
 private:
@@ -29,13 +29,13 @@ private:
     // avoiding the static initialization order fiasco
     static std::vector<std::string>& get_filenames();
     static std::unordered_map<std::string, size_t>& get_filename_map();
-    
+
     // Get or create filename index
     static size_t get_filename_index(const std::string& filename);
-    
+
     // Get filename from index (empty string for index 0)
     static const std::string& get_filename_by_index(size_t index);
-    
+
     size_t filename_index_ = 0;  // Index into filenames vector (0 = empty)
-    int line_num_ = 0;
+    size_t line_num_ = 0;
 };
